requirements.txt
annotated-types==0.7.0
anyio==4.6.2.post1
Authlib==1.3.1
blinker==1.9.0
cachetools==5.5.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
cryptography==43.0.3
distro==1.9.0
exceptiongroup==1.2.2
Flask==3.1.0
Flask-Cors==5.0.0
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.2
google-auth==2.36.0
google-auth-oauthlib==1.2.1
greenlet==3.1.1
grpcio==1.68.0
grpcio-health-checking==1.68.0
grpcio-tools==1.68.0
gunicorn==21.2.0
gspread==6.1.4
h11==0.14.0
httpcore==1.0.6
httpx==0.27.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.4
jiter==0.7.0
MarkupSafe==3.0.2
numpy==2.1.3
oauthlib==3.2.2
openai==1.53.0
pandas==2.2.3
protobuf==5.28.3
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pydantic==2.9.2
pydantic_core==2.23.4
pyee==12.0.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.36
tqdm==4.66.6
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
validators==0.34.0
weaviate-client==4.9.6
Werkzeug==3.1.3
WTForms==3.2.1